requirements.txt
alembic==1.6.5
appnope==0.1.2
astroid==2.5.6
autopep8==1.5.7
backcall==0.2.0
boto3==1.20.11
botocore==1.23.11
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cycler==0.10.0
decorator==5.0.9
Faker==8.5.1
Flask==2.0.1
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
idna==2.10
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
isort==5.8.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
jupyter-client==6.1.12
jupyter-core==4.7.1
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
Mako==1.1.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mccabe==0.6.1
nose==1.3.7
numpy==1.21.0
pandas==1.2.5
parso==0.8.2
peewee==3.14.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pip==21.1.2
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pycodestyle==2.7.0
Pygments==2.9.0
pylint==2.8.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.19.2
python-editor==1.0.4
pytz==2021.1
pyzmq==22.1.0
requests==2.25.1
s3transfer==0.5.0
setuptools==49.2.1
six==1.16.0
SQLAlchemy==1.4.17
sqlparse==0.4.1
tabulate==0.8.9
text-unidecode==1.3
toml==0.10.2
tornado==6.1
traitlets==5.0.5
urllib3==1.26.5
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
yoyo-migrations==7.3.1